Cleanup a bunch of code.
Use Glib macros for defining a type, saves a bunch
of code and is thread-safe.

Drop ; after G_{BEGIN,END}_DECLS macros.

Move G_GNUC_INTERNAL macros in front of the function
to avoid problems with Solaris. Remove them from the
code that is compiled as an executable.

Tabs and spaces cleanups.


(Old svn revision: 30340)