Make ExoLazyCellRenderer available, suggested in bug #1266 in 2005.
This new ExoLazyCellRenderer interface adds a new method
render_and_resize() to cell renderers implementing this interface. This
can be used for applications which load icons only on-demand (like
Thunar).

It allows cell renderers to modify the size of the rendered cells before
rendering them and pass the modifications back to the icon view.

This implementation still has a few resizing issues but maybe we can
sort them out one day...

The original code was written by Benny, I've modified the patch to apply
against the current development branch.