Do not insert symlink icons in the model.
When loading the icons, check if the filename of the icon
is a symlink to another icon in the theme. If so merge the
"alternative" name in the real icon. This saves a lot of
icon duplicates in the icon chooser, while they are still
selected because the alternative names are still known.

This add some complexity, but might be a lot faster in the
end, because on normal themes around 50% of the icons are
merged, so a lot less pixbuf loading/rendering.