Don't quote arguments passed to terminal emulators (bug 10731)
Commit 8cd3233 added argument quoting to several helpers to properly
pass URIs that contain spaces to web browsers like Firefox.

This broke invocations such as the following:

  $ exo-open --launch TerminalEmulator screen -S test

resulting in errors similar to:

  Failed to execute child
  Failed to execute child process "screen -S test" (No such file or
  directory)

Fix this by reverting the quoting changes made to TerminalEmulator
helpers. The Sakura terminal, surprisingly enough, seems to require
quotes around the command-line passed to it so I kept it with quotes.

Signed-off-by: Sean Davis <smd.seandavis@gmail.com>