build: clang: Silence -Wcast-align
Since ExoBinding is now deprecated, there is no point in changing the
code to fix these warnings.

For GdkPixbuf it seems to be false positives, see
https://gitlab.xfce.org/xfce/exo/-/issues/67#note_50795

For ExoIconView it could be improved, but I don't want to bother with
that now and risk introducing regressions. I'm going to open a new
dedicated issue.

Fixes: #67